Struct tartan_pci::config::BaseAddressRegister

#[repr(transparent)]
pub struct BaseAddressRegister(_);

A pointer to a memory or I/O space that is used to interact with the function.

The meaning of the addressed space is defined by the specific function’s interface.

Implementations

impl BaseAddressRegister

pub fn address_space(self) -> AddressSpace

Indicates whether this BAR points to memory or I/O space

pub fn prefetchable(self) -> bool

Indicates whether the host can safely cache the memory addressed by this BAR

pub fn address_width(self) -> AddressWidth

Size of address bus supported by this PCI function.

Addresses in I/O space are always 32-bits wide.

pub fn address_u32(self) -> u32

The address as a 32-bit number.

If the address_width for this BAR is 64 bits, then this will only return the lower half of the address.

pub fn address_u64(self, next_register: u32) -> u64

The address as a 64-bit number, when combined with the value of the following register.

If the address_width of this BAR is 32 bits, then the value of the next register will be ignored. Thus, this is safe to use with any address width.