Struct tartan_pci::config::CommandRegister

pub struct CommandRegister(/* private fields */);

Control bits set by the host

Implementations

impl CommandRegister

pub fn interrupt_disabled(&self) -> bool

Controls whether the device is prohibited from asserting its interrupt pin. Does not apply to MSI interrupts.

pub fn fast_back_to_back_enabled(&self) -> bool

Controls whether the device may use fast back-to-back transactions to multiple devices when acting as a bus master.

pub fn system_error_enabled(&self) -> bool

Controls whether the device may signal critical errors in addresses or special cycle operations.

pub fn parity_error_response(&self) -> bool

Controls whether the device responds normally to parity errors (true) or whether it must simply set the parity error bit in the status register and ignore it (false).

pub fn vga_palette_snoop(&self) -> bool

Controls whether a VGA or graphics card device may “snoop” on VGA palette buffer writes sent to a different device.

pub fn write_and_invalidate_enable(&self) -> bool

Controls whether the device may send a “Memory Write and Invalidate” command when acting as a bus master.

pub fn special_cycle(&self) -> bool

Controls whether the device may respond to “special cycle” messages, used to communicate system state to the device (e.g., shutting down, halted)

pub fn bus_master(&self) -> bool

Controls whether the device may act as a master of the PCI bus

pub fn memory_space(&self) -> bool

Controls whether the device may respond to accesses through memory space

pub fn io_space(&self) -> bool

Controls whether the device may respond to accesses through I/O space

pub fn set_interrupt_disabled(&mut self, value: bool)

Controls whether the device is prohibited from asserting its interrupt pin. Does not apply to MSI interrupts.

pub fn with_interrupt_disabled(&mut self, value: bool) -> Self

Controls whether the device is prohibited from asserting its interrupt pin. Does not apply to MSI interrupts.

pub fn set_fast_back_to_back_enabled(&mut self, value: bool)

Controls whether the device may use fast back-to-back transactions to multiple devices when acting as a bus master.

pub fn with_fast_back_to_back_enabled(&mut self, value: bool) -> Self

Controls whether the device may use fast back-to-back transactions to multiple devices when acting as a bus master.

pub fn set_system_error_enabled(&mut self, value: bool)

Controls whether the device may signal critical errors in addresses or special cycle operations.

pub fn with_system_error_enabled(&mut self, value: bool) -> Self

Controls whether the device may signal critical errors in addresses or special cycle operations.

pub fn set_parity_error_response(&mut self, value: bool)

Controls whether the device responds normally to parity errors (true) or whether it must simply set the parity error bit in the status register and ignore it (false).

pub fn with_parity_error_response(&mut self, value: bool) -> Self

Controls whether the device responds normally to parity errors (true) or whether it must simply set the parity error bit in the status register and ignore it (false).

pub fn set_vga_palette_snoop(&mut self, value: bool)

Controls whether a VGA or graphics card device may “snoop” on VGA palette buffer writes sent to a different device.

pub fn with_vga_palette_snoop(&mut self, value: bool) -> Self

Controls whether a VGA or graphics card device may “snoop” on VGA palette buffer writes sent to a different device.

pub fn set_write_and_invalidate_enable(&mut self, value: bool)

Controls whether the device may send a “Memory Write and Invalidate” command when acting as a bus master.

pub fn with_write_and_invalidate_enable(&mut self, value: bool) -> Self

Controls whether the device may send a “Memory Write and Invalidate” command when acting as a bus master.

pub fn set_special_cycle(&mut self, value: bool)

Controls whether the device may respond to “special cycle” messages, used to communicate system state to the device (e.g., shutting down, halted)

pub fn with_special_cycle(&mut self, value: bool) -> Self

Controls whether the device may respond to “special cycle” messages, used to communicate system state to the device (e.g., shutting down, halted)

pub fn set_bus_master(&mut self, value: bool)

Controls whether the device may act as a master of the PCI bus

pub fn with_bus_master(&mut self, value: bool) -> Self

Controls whether the device may act as a master of the PCI bus

pub fn set_memory_space(&mut self, value: bool)

Controls whether the device may respond to accesses through memory space

pub fn with_memory_space(&mut self, value: bool) -> Self

Controls whether the device may respond to accesses through memory space

pub fn set_io_space(&mut self, value: bool)

Controls whether the device may respond to accesses through I/O space

pub fn with_io_space(&mut self, value: bool) -> Self

Controls whether the device may respond to accesses through I/O space