Struct tartan_pci::config::HeaderRegister3

pub struct HeaderRegister3(/* private fields */);

The fourth register of the standard configuration header for all PCI devices.

Implementations

impl HeaderRegister3

pub fn self_test(&self) -> SelfTest

Control and status flags for a function’s self test capability, if present.

pub fn multi_function(&self) -> bool

If true, then this device responds to at least some function numbers >= 1. The flag is only meaningful on function zero.

pub fn header_type(&self) -> u8

Indicates which of the standard PCI configuration header formats that this function supports, and indirectly, whether this function is a PCI or CardBus bridge.

Currently, the defined values are:

0x00 - general purpose

0x01 - PCI bridge

0x02 - PCI-to-CardBus bridge

pub fn latency_timer(&self) -> u8

The value of the bus master’s latency timer in PCI clock cycles

pub fn cache_line_size(&self) -> u8

Informs the device of the size of the system’s cache lines, in 32-bit units.

pub fn set_self_test(&mut self, value: SelfTest)

Control and status flags for a function’s self test capability, if present.

pub fn with_self_test(&self, value: SelfTest) -> Self

Control and status flags for a function’s self test capability, if present.

pub fn set_multi_function(&mut self, value: bool)

If true, then this device responds to at least some function numbers >= 1. The flag is only meaningful on function zero.

pub fn with_multi_function(&mut self, value: bool) -> Self

If true, then this device responds to at least some function numbers >= 1. The flag is only meaningful on function zero.

pub fn set_header_type(&mut self, value: u8)

Indicates which of the standard PCI configuration header formats that this function supports, and indirectly, whether this function is a PCI or CardBus bridge.

Currently, the defined values are:

0x00 - general purpose

0x01 - PCI bridge

0x02 - PCI-to-CardBus bridge

pub fn with_header_type(&self, value: u8) -> Self

Indicates which of the standard PCI configuration header formats that this function supports, and indirectly, whether this function is a PCI or CardBus bridge.

Currently, the defined values are:

0x00 - general purpose

0x01 - PCI bridge

0x02 - PCI-to-CardBus bridge

pub fn set_latency_timer(&mut self, value: u8)

The value of the bus master’s latency timer in PCI clock cycles

pub fn with_latency_timer(&self, value: u8) -> Self

The value of the bus master’s latency timer in PCI clock cycles

pub fn set_cache_line_size(&mut self, value: u8)

Informs the device of the size of the system’s cache lines, in 32-bit units.

pub fn with_cache_line_size(&self, value: u8) -> Self

Informs the device of the size of the system’s cache lines, in 32-bit units.