tartan_pci::config

Struct StatusRegister

pub struct StatusRegister(/* private fields */);

Status bits indicated by the function

Implementations

impl StatusRegister

pub fn parity_error_detected(&self) -> bool

Indicates taht the device has detected a parity error, even if it is ignored.

pub fn system_error_signaled(&self) -> bool

Indicates that the device has detected a critical error in an address or special cycle operation.

pub fn master_abort_received(&self) -> bool

Indicates whether the transaction was aborted with a Target-Abort while this device was acting as the bus master.

pub fn target_abort_received(&self) -> bool

Indicates whether the transaction was aborted with a Target-Abort.

pub fn target_abort_signaled(&self) -> bool

Indicates that this function generated an abort on the transaction (as opposed to a master abort).

pub fn device_select_timing(&self) -> u8

Indicates how quickly the device can assert the DEVSEL# pin.

Valid values are:

0x0 - fast

0x1 - medium

0x2 - slow

pub fn master_parity_error(&self) -> bool

Indicates whether a partity error was detected while this device was acting as the bus master.

pub fn fast_back_to_back_capable(&self) -> bool

Indicates whether the device supports receiving back-to-back transactions that are targeted at multiple devices.

pub fn double_clock_capable(&self) -> bool

Indicates whether the device can run at 66 MHz (true).

pub fn capabilities_list_available(&self) -> bool

Indicates whether the configuration space contains a linked list of capabilities, beginning at the offset listed in Type0HeaderRegister13.

pub fn interrupt_status(&self) -> bool

Indicates whether the device wants to signal an interrupt, even if it is prohibited from asserting its interrupt pin.

pub fn set_parity_error_detected(&mut self, value: bool)

Indicates taht the device has detected a parity error, even if it is ignored.

pub fn with_parity_error_detected(&mut self, value: bool) -> Self

Indicates taht the device has detected a parity error, even if it is ignored.

pub fn set_system_error_signaled(&mut self, value: bool)

Indicates that the device has detected a critical error in an address or special cycle operation.

pub fn with_system_error_signaled(&mut self, value: bool) -> Self

Indicates that the device has detected a critical error in an address or special cycle operation.

pub fn set_master_abort_received(&mut self, value: bool)

Indicates whether the transaction was aborted with a Target-Abort while this device was acting as the bus master.

pub fn with_master_abort_received(&mut self, value: bool) -> Self

Indicates whether the transaction was aborted with a Target-Abort while this device was acting as the bus master.

pub fn set_target_abort_received(&mut self, value: bool)

Indicates whether the transaction was aborted with a Target-Abort.

pub fn with_target_abort_received(&mut self, value: bool) -> Self

Indicates whether the transaction was aborted with a Target-Abort.

pub fn set_target_abort_signaled(&mut self, value: bool)

Indicates that this function generated an abort on the transaction (as opposed to a master abort).

pub fn with_target_abort_signaled(&mut self, value: bool) -> Self

Indicates that this function generated an abort on the transaction (as opposed to a master abort).

pub fn set_device_select_timing(&mut self, value: u8)

Indicates how quickly the device can assert the DEVSEL# pin.

Valid values are:

0x0 - fast

0x1 - medium

0x2 - slow

pub fn with_device_select_timing(&self, value: u8) -> Self

Indicates how quickly the device can assert the DEVSEL# pin.

Valid values are:

0x0 - fast

0x1 - medium

0x2 - slow

pub fn set_master_parity_error(&mut self, value: bool)

Indicates whether a partity error was detected while this device was acting as the bus master.

pub fn with_master_parity_error(&mut self, value: bool) -> Self

Indicates whether a partity error was detected while this device was acting as the bus master.

pub fn set_fast_back_to_back_capable(&mut self, value: bool)

Indicates whether the device supports receiving back-to-back transactions that are targeted at multiple devices.

pub fn with_fast_back_to_back_capable(&mut self, value: bool) -> Self

Indicates whether the device supports receiving back-to-back transactions that are targeted at multiple devices.

pub fn set_double_clock_capable(&mut self, value: bool)

Indicates whether the device can run at 66 MHz (true).

pub fn with_double_clock_capable(&mut self, value: bool) -> Self

Indicates whether the device can run at 66 MHz (true).

pub fn set_capabilities_list_available(&mut self, value: bool)

Indicates whether the configuration space contains a linked list of capabilities, beginning at the offset listed in Type0HeaderRegister13.

pub fn with_capabilities_list_available(&mut self, value: bool) -> Self

Indicates whether the configuration space contains a linked list of capabilities, beginning at the offset listed in Type0HeaderRegister13.

pub fn set_interrupt_status(&mut self, value: bool)

Indicates whether the device wants to signal an interrupt, even if it is prohibited from asserting its interrupt pin.

pub fn with_interrupt_status(&mut self, value: bool) -> Self

Indicates whether the device wants to signal an interrupt, even if it is prohibited from asserting its interrupt pin.