Crate tartan_pci

Support for Peripheral Component Interconnect (PCI) and PCI Express (PCIe) devices.

Modules

access
Access methods for PCI configuration space.

config
Data exposed in PCI configuration space.

Structs

CapabilityEntry
Information to identify and locate capability registers

Constants

INVALID_VENDOR
Placeholder value that will always be returned in the vendor field when querying PCI configuration space for a device that does not exist.

MAX_DEVICE
Highest device number allowed by the PCI specification.

MAX_FUNCTION
Highest function number allowed by the PCI specification.

Functions

check_valid
Return true if a function is available at the specified selector.

enumerate_bus
Iterate over the devices and functions present on the specified bus.

enumerate_bus_devices
Iterate over the devices present on the specified bus.

enumerate_device_functions
Iterate over the functions available on the specified device.

iter_capabilities
Iterate over all capabilities defined in the configuration space for a PCI function