Trait tartan_pci::access::ConfigAccess

pub trait ConfigAccess {
    fn get_register(&self, selector: ConfigSelector, register: u16) -> u32;

    fn set_register(&self, selector: ConfigSelector, register: u16, value: u32);

    fn get_fixed_register<T: FixedConfigRegister>(
        &self, 
        selector: ConfigSelector
    ) -> T { ... }

    fn set_fixed_register<T: FixedConfigRegister>(
        &self, 
        selector: ConfigSelector, 
        value: T
    ) { ... }
}

Provides methods to access PCI configuration space.

Implementations may only offer access to a limited number of registers on a specific range of segments/busses.

Required methods

fn get_register(&self, selector: ConfigSelector, register: u16) -> u32

Get a specific 32-bit register in PCI configuration space.

fn set_register(&self, selector: ConfigSelector, register: u16, value: u32)

Set a specific 32-bit register in PCI configuration space.

Loading content...
Provided methods

fn get_fixed_register<T: FixedConfigRegister>(
&self,
selector: ConfigSelector
) -> T

Get a register in PCI configuration space with an offset defined by the output type.

fn set_fixed_register<T: FixedConfigRegister>(
&self,
selector: ConfigSelector,
value: T
)

Set a register in PCI configuration space with an offset defined by the input type.

Loading content...