Module tartan_pci::config

Data exposed in PCI configuration space.

Structs

BaseAddressRegister
A pointer to a memory or I/O space that is used to interact with the function.

CommandRegister
Control bits set by the host

GenericCapabilityRegister
Structure shared by all capability registers

HeaderRegister0
The first register of the standard configuration header for all PCI devices.

HeaderRegister1
The second register of the standard configuration header for all PCI devices.

HeaderRegister2
The third register of the standard configuration header for all PCI devices.

HeaderRegister3
The fourth register of the standard configuration header for all PCI devices.

SelfTest
Control and status flags for a function's self test capability, if present.

StatusRegister
Status bits indicated by the function

Type0HeaderRegister11
Twelfth register of the standard configuration header for general-purpose devices (header type 0x00).

Type0HeaderRegister13
Fourtheenth register of the standard configuration header for general-purpose devices.

Type0HeaderRegister15
Sixteenth register of the standard configuration header for general-purpose devices.

Enums

AddressSpace
Differentiates between memory and I/O addresses

AddressWidth
Size of address bus supported by a PCI function

Traits

FixedConfigRegister
A register type that is always located at a specific offset